Extract poll renderer component and fix styling
